moved Objects to separate Class
